requirements.txt
beautifulsoup4==4.4.1
boto==2.48.0
certifi==2017.11.5
chardet==3.0.4
click==6.6
dj-database-url==0.4.1
Django==1.9.6
django-appconf==1.0.2
django-compressor==2.2
django-filter==0.13.0
django-htmlmin==0.9.1
django-mailgun==0.9.1
django-paypal==0.3.2
django-storages==1.6.5
djangorestframework==3.3.3
gunicorn==19.6.0
html5lib==0.9999999
idna==2.6
Markdown==2.6.6
ordereddict==1.1
pep8==1.7.0
Pillow==3.2.0
psycopg2==2.6.1
pylibmc==1.5.1
python-dotenv==0.5.0
pytz==2016.4
qrcode==5.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
simplejson==3.8.2
six==1.11.0
sorl-thumbnail==12.3
urllib3==1.22
